This is an experiment.
Contributed software, like the timezone database, is always imported,
often a merge-from-vendor is done and most of the times it is MFCd
into the latest branch and the branch before that.

Updating the release note documentation is a thing which is sometimes
done, sometimes done by a third person and sometimes not done at
all.

By creating a datafile to store the information about the software,
its versions, its import dates, its MFV dates and its MFC dates, I
hope that it will be easier for commiters (lazy bums as they are!)
and re@ to make sure that the information about these updates is
tracked and easy to be used to generate release notes.

First steps:
Create a copy of the release directories for the 6.x and higher
releases and branches, adjust the article.sgml files 7.x and higher
releases and branches.
Create the right catalog for the contributed software entity.
Create the configuration file and the beginning of a script to
convert the configuration into a piece of information which can be
used by the releasenote processing software.