Remove assumptions about the max # channels: all channel list data structures are now allocated to hold MAXCHAN entries (1536). This, for example, allows the kernel to return > IEEE80211_CHAN_MAX entries for calls like IEEE80211_IOC_DEVCAPS.
Note that with the size of struct ieee80211_channel now 16 bytes
1.5K entries is near the max we can pass through the ioctl api's
because i_len is 16-bits.  However that should be sufficient for
the forseeable future as the max channel list that currently is
passed back has 550 entries and before that grows much more we're
likely to switch to an alternate scheme for passing info back+forth.