Fix the NF calibration to only occur if a periodic calibration is not in progress.
Since I've also seen periodic calibrations take forever (as in, up
to an hour to do the 64 samples needed for a percal), longcal may
never get set during a time where there isn't an in-progress percal.

To avoid this, perform a longcal immediately after the last successful
periodic calibration sample is read. There shouldn't be any further
periodic calibrations performed if *isCalDone is set to AH_TRUE.
This means that there'll be at least one NF calibration and sample
done at the end of each periodic calibration.

I'm not sure whether a NF sample and/or calibration can be taken in
between calibration samples (whether it'll upset the calibration
process, which I don't have any documentation on atm) so I'm not
interleaving NF and periodic calibration sampling.