Begin making ampdu tx work (but only with non-aggregate packets.)
* the TID lookup using ieee80211_gettid() was wrong, instead use the
  AC from the mbuf and convert that to a TID.
* Since A-MPDU TX packets aren't supposed to be stuffed on the cabq
  when the station is in PS mode, don't do that. Instead, we'll eventually
  handle this lter.
* Move the eventual TXQ selection out to ath_tx_start().
* write a function to assign seqno's to A-MPDU TX frames.

This works enough to TX non-aggregate frame in A-MPDU but it's missing
a lot of support. In particular, BAR's aren't sent when the window needs
shifting (because of a TX failure) and there's currently no logic to
only queue frames which fit inside the BA window.

There's also a problem with the initial sequence numbers used just after
the ADDBA exchange, so there's a (large) BA window shift. That needs to be
fixed.