Begin fleshing out the BAR handling.
net80211 implements BAR frame queuing and completion handofing.
The first place a BAR should be TXed is when the software retry
limit has been exceeded for a TID currently doing aggregation.
The process should be (just so I don't forget):

* pause the TID;
* send the BAR;
* unpause the TID in ath_bar_response()

If multiple BAR's are sent, paused will be incremented and decremented
appropriately until all BAR frames have been transmitted (succesfully
or otherwise.)

However, using it with the current code is likely not going to work
because the ath TX code isn't re-entrant w/ the way locking currently
is implement.