Fix some BAW brokenness introduced by previous commits.
The BAW tracking was completely broken. ath_tx_swq() was blanking the BAW
flag in the ath_buf, thus no frames were ever thrown into the BAW tracking
code.

This restores the BAW tracking functionality but there's again a race between
the pre-ADDBA sequence numbers and the post-ADDBA BAW. I'll look into this
and fix it in a subsequent commit.