As much as I dislike it, migrate the driver back to doing direct dispatch to hardware.
This matches what the Linux/atheros reference code does.

If the hardware txq isn't busy, start filling it with non-aggregate
frames. If it's sufficiently busy, start aggregating frames in the
background.

The aim is to massage this code back to match the structure of the
reference code in order to make porting/debugging easier. I may
move it back to task-based TX at a later stage, but only once the
rest of the TX path has been fully debugged.

This drops the CPU use a little but there's still a problem breaking
> 100mbit on the MIPS dev boards I'm using. (The same dev board
can do 250mbit RX on the same NIC, so it's not necessarily unable
to do it.)