Don't clear the MRR fields for aggregate frames.
This allows for faster recovery if higher MCS rates begin to fail,
as lower MCS rates are now tried.

There are still issues forming large aggregates because of the logic
involved in selecting the largest aggregate frame size (which is
based on the lowest 4ms frame length value in the rate series),
but that's a later issue to solve.