Begin adding some further debugging to chase down the TX hangs that I've seen under TCP TX load.
It looks like the underlying problem is that some packets that are -before-
the BAW are making it onto the software TX queue. These frames are outside
the BAW (but before it, not after it) and thus the TX scheduling stops.

The source of those grossly out of order (and likely already successfully
TXed, but that's currently unknown) is currently unknown.