Add a function rtlookup() that copies out the relevant information from an rtentry instead of returning the rtentry.  This avoids the need to lock the rtentry and to increase the refcount on it.
Convert ip_output() to use rtlookup() in a simplistic way.  Certain
seldom used functionality may not work anymore and the flowtable
isn't available at the moment.

Convert ip_fastfwd() to use rtlookup().

This code is meant to be used for profiling and to be experimented
with further to determine which locking strategy returns the best
results.