First commit that concerns event aggregation. Augment the callout structure so that consumers other than actual time at which callout should fire may specify a tolerance interval.
Rather than looking for the next callout event in callout_tick()
determine a range [t-delta;t'+delta'] deriving it from the tolerance
parameter specified by clients so that's suitable for a given number of events,
and schedule an interrupt in the middle of such range.

Add some comments to better specify what we're doing.

Suggested by:		mav
Reviewed by:		mav